pyproject.toml
[project]
name = "api-monitoring-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.2",
    "django>=6.0.3",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "gunicorn>=25.1.0",
    "psycopg-binary>=3.3.3",
    "python-dotenv>=1.2.2",
    "redis>=7.2.1",
]
